components/software/ground_station/pyproject.toml
[project]
name = "ground-station"
version = "0.1.0"
description = "Ground station waterfall viewer for ADALM-Pluto SDR"
requires-python = ">=3.11"
dependencies = [
    "pyside6",
    "numpy",
    "pyadi-iio",
]

[project.scripts]
ground-station = "ground_station.app:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
